fixed64: strict-clean bitmap/actor/vectors/sbarinfo/gl_material/farchive headers
Behavior-preserving conversions up the header graph:
  - bitmap.h: pixel-blend results cast (int) to the BYTE dest / MIN<int>
    arg (0..255 values, same truncation as before).
  - actor.h: bob finesine index cast (int) (masked to 0..8191).
  - sbarinfo.h: getAlpha default arg (int)FRACUNIT.
  - gl_material.h: GetScaled*Offset return (int)DivScale16(...).
  - vectors.h: TVector3::operator- routes through double(), which is what
    the sole (double,double,double) ctor already did for TVector3<fixed_t>.
  - farchive.h: add an FArchive operator<< for the strong Fixed that
    serializes the raw 48.16 value (same on-disk layout as SQWORD; SAVEVER
    already bumped).

Flag-off engine build green; tests green. Strict unique sites 1402 -> 1152.

+++ src/zandronum/src/g_shared/sbarinfo.h
@@ -77,7 +77,7 @@ struct Popup
 	bool isDoneMoving();
 	int getXOffset();
 	int getYOffset();
-	int getAlpha(int maxAlpha=FRACUNIT);
+	int getAlpha(int maxAlpha=(int)FRACUNIT);
 	int getXDisplacement();
 	int getYDisplacement();
 };
